Updated Objective-C sources to make the project buildable for arm64.

+++ Objective-C/iMMPI/iMMPI/AnalysisViewController.m
@@ -477,7 +477,7 @@ didSelectRowAtIndexPath: (NSIndexPath *) indexPath
     NSUInteger index = [group indexForRecord: self.record];
     
     if (index > 0)
-        cell.indexLabel.text = [NSString stringWithFormat: @"%d.", index];
+        cell.indexLabel.text = [NSString stringWithFormat: @"%ld.", (long)index];
     else
         cell.indexLabel.text = nil;
